requirements_dev.txt
attrs==22.2.0
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pycallgraph==1.0.1
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
tomli==2.0.1
tqdm==4.65.0
